Move cookie helper outside `pages`
