+++ packages/app-project/pages/_app.js
@@ -8,8 +8,8 @@ import React from 'react'
 import { createGlobalStyle } from 'styled-components'
 import UrlParse from 'url-parse'
 
-import { getCookie } from './helpers/cookies'
 import AuthModal from '../src/components/AuthModal'
+import getCookie from '../src/helpers/getCookie'
 import GrommetWrapper from '../src/helpers/GrommetWrapper'
 import Head from '../src/components/Head'
 import ProjectHeader from '../src/components/ProjectHeader'
