+++ packages/app-project/src/helpers/getCookie/getCookie.js
@@ -1,6 +1,6 @@
 import cookie from 'cookie'
 
-export function getCookie (context, name) {
+export default function getCookie (context, name) {
   if (context && context.req && context.req.headers.cookie) {
     const parsedCookie = cookie.parse(context.req.headers.cookie)
     if (parsedCookie && parsedCookie[name]) {
