+++ packages/app-project/src/helpers/getCookie/index.js
@@ -1,0 +1,1 @@
+export { default } from './getCookie'
